Basic implementation of Valyrian's Crew (#2136)
* Add support for card counts in test deck builder

Previously, if you wanted to add a specific number of copies of a single
card to a deck in a test, you would have to list that card's name that
many times in the call to `buildDeck`. Now there's a short hand
available that allows you to specify the count directly.

Example:
```javascript
this.buildDeck('stark', [
  'Trading with the Pentoshi', 'A Noble Cause',
  'Nightmares', 'Heart Tree Grove', 'Needle',
  { name: 'Arya Stark (Core)', count: 3 }
]);
```

* Facedown cards should be visible to controller

* Specify UUID on facedown cards

Previously, it was unnecessary to identify facedown cards because
players could not interact with them. With Valyrian's Crew, however, the
facedown cards become attachments that players can target with
abilities, thus needing a unique identifier.

* Ensure onCardMoved is raised for attachments

Previously, if a card was attached to another card and then detached
from that card, `onCardMoved` would not be raised properly. This could
lead to effects not being properly recalculated on the attachment.

* Use printed type for ability triggering

Due to event cards being able to be triggered while facedown under
Valyrian's Crew, the printed type needs to be checked in most places
because the type of the card will be an attachment.

* Convert `canPlay` effect to take predicate

Previously the `canPlay` effect could only take individual cards. It can
now take a predicate function that matches cards e.g. in a certain
location.

* Only allow cards w/ cost for Trading w/ Qohor

Facedown cards and cards with a '-' cost should be excluded from
eligibility for Trading with Qohor.

* Allow facedown attachments + actions from them

* Implement Valyrian's Crew

Adds basic effect for being able to marshal facedown cards.