Check maximums for non-event abilities (#2167)
Previously, only abilities on event cards were being checked for
maximums, and it was being done as a cost. This led to a bug where the
max on character / attachment abilities would not be properly enforced.
The timing would also have been incorrect in the case of the cost being
cancelled (Drowned God Fanatic).

Now, maximums are enforced when requirements for the ability are
checked, and marked as used at the same time as limits. This should
ensure that Drowned God Fanatic is properly limited to once per round.