Allow search prompts to use the draw deck (#37)
* Extract draw deck to separate component

When considering the menu that appears when clicking the draw pile and
the "Close and Shuffle" button on the popup, there was enough behavior
on the draw deck to extract it as a separate component.

* Send draw deck visibility to the server

Previously, draw deck visibility was toggled on the server by tracking
when the pile was opened and closed via player clicks. However, this
didn't account for the popup opening and closing without user input due
to search prompts. This is now handled through an onPopupChange event
that is fired whenever the popup visibility has changed - either through
users clicking or from prompts activating. The new visibility is then
sent to the server to update viewing state.

* Initialize popup by card selectability

For House with the Red Door, the draw popup needs to be open as soon as
the game starts. Thus, the `showPopup` flag needs to be set based on
whether there are any selectable cards during the constructor.

* Hide "Close and Shuffle" button during deck search

* Enable opponent draw popup when cards are present

When you reveal Exchange of Information, the opponent will receive the
top 10 cards in your draw deck. During this and similar searches, your
draw deck should be able to automatically open for the opponent, as well
as the opponent to toggle the popup freely.