Allow dupes when card abilities prevented
Using a dupe to save is considered a "game ability" instead of a "card
ability." Since manual dupes are set up as using the ability system,
they were improperly considered a "card ability" when cards like
Winterfell, Core Catelyn, or King in the North were in play.

Now, abilities can be defined as a game ability and thus skip the checks
involved with effects preventing card abilities.