Allow ability messages defined outside of handler
Adds a `message` property to triggered abilities in order to allow
messages to be output before the handler for the ability executes. This
will allow messages to be output to the game log before players are
prompted for cancel abilities (Treachery, Hand's Judgment, etc). It will
also enable the game action mechanism in the future, which will require
removing / replacing the existing `handler` property for abilities.