Remove menus for opponent actions
The menu will no longer display when you click a card with an action
when the opponent controls that card. For example, using Confiscation to
discard a Seal of the Hand controlled by an opponent won't bring up the
"Stand attached character" menu. If the action is allowed to be
triggered by any player (e.g. Bronn), the menu will display for all
appropriate players.