Send 'X' strength to client (#2221)
Previously, if an X strength character reached 0 STR (e.g. a Stark-less
Dacey Mormonet or Beric Dondarrion with no kiss tokens), their strength
wouldn't be displayed by the client. This is because the client was
checking the numerical printed value of strength against the current
strength. For X STR characters, that's 0.

By sending 'X' STR to the client directly instead of 0, the client can
decide whether it should display a 0 STR value.