Format arrays of players without using avatars (#2235)
The Free Folk agenda was displaying the player receiving claim using
their avatar, like a chat message from that player. This is because the
`Claim.recipients` field is an array in case of Master of Whispers in
melee, and player objects in arrays used the avatar. Now, only chat
messages are explicitly formatted to display with avatars.