Anonymous spectators are spectators
When an error occurs, the current game state is serialized for the error
report as if an anonymous spectator were watching the game. However, the
spectator detection changed in #2228 was not applied to the
AnonymousSpectator class, causing a crash when an error report would
occur.

Fixes THRONETEKI-2BB
Fixes THRONETEKI-2BC