Warn when a player takes an action out of turn
Previously, a warning message was added when a player takes an action
outside of a legal action window. However, while an action window was
open, each player could take actions no matter whose turn it was to take
an action.

Now, players can still take an action outside of their turn during an
action window, but it will add a warning message to the chat log
reminding the player that it was out of turn. This will keep play
unrestrictive while also making clear a game rule has been violated.