Implement /cancel-challenge command
When a player uses the /cancel-challenge command, each opponent will be
prompted whether they want to allow the cancellation. If all players
agree, the challenge will be cancelled and untracked. Any abilities used
before the cancellation will need to be undone manually by the players.

If a single player does not want to allow the cancellation, the
challenge continues as normal.