Implement /cancel-challenge (#2280)
* Allow challenges to be untracked

When a challenge is untracked, it removes it from the list of initiated
challenges as well restores the ability for the player to initiate that
challenge again.

* Allow game pipelines to be cleared

Previously, only individual steps in a pipeline could be cancelled.
However, for the /cancel-challenge command it will be necessary to clear
out the entire pipeline under the ChallengeFlow sub-pipeline.

* Add onCompleted handler to UiPrompt

When a UiPrompt is completed, it will call the onCompleted handler that
inheritors may implement. This method allows execution of code for the
prompt after all player choices have been made.

* Implement /cancel-challenge command

When a player uses the /cancel-challenge command, each opponent will be
prompted whether they want to allow the cancellation. If all players
agree, the challenge will be cancelled and untracked. Any abilities used
before the cancellation will need to be undone manually by the players.

If a single player does not want to allow the cancellation, the
challenge continues as normal.