Fix Seasoned Woodsman (#2283)
* Validate event properties for overlapping params

Certain events were using `parent` as a parameter name, which was
silently overwritten by the event object internally. Now, if an event
param has the same key as something used by the event object internally,
an exception is raised.

* Fix onCardAttached event + Seasoned Woodsman

When the onCardAttached event was modified to add support for
attachments coming out of the shadows, it gained a parent event. This
conflicted with the `parent` property passed in the event parameters,
causing Seasoned Woodsman to stop triggering.