Prompt all players for actions when setting is on (#2285)
Previously, when a single player had an action window setting turned on,
only the player with the setting turned on would be prompted. Any player
without the setting turned on would auto-pass when it was their turn in
the action window, unless another player had actually taken an action.

This could lead to confusion for players that do not understand that
they will auto-pass without the option turned on. A typical scenario
would be first player has the window off, and thus auto-passes. Then,
the second player has the window on, and uses an action. Then, the first
player, not realizing they auto-passed, complains that they should have
had the first opportunity to use an action.

To alleviate this confusion, auto-pass has been removed unless all
players have the action window turned off. If any player has the setting
turned on, each player will be prompted in first-player order, even if
they personally have the setting turned off.