Clear tokens when transitioning locations
Previously, tokens and power were only cleared when a card left play.
However, now tokens are able to be placed on cards in the shadows, which
need to be discarded when they leave shadows.

Now, tokens and power are cleared when a card is moved between areas. If
it is moved within the same area (e.g. an attachment being moved from
one card to another), the tokens remain.