Enable hover based on whether card code is present (#45)
Previously, squishable card panels handled card hover using the `isMe`
property. Meereen creates a squishable card panel but does not have
access to the property, and was thus not allowing the controller to look
at the cards underneath.

Now, card hover is controlled by the presence of the card code. The
server will not send the card code unless it has already determined that
the player can see the card. Thus we no longer need to rely on the
`isMe` property for such visibility.