Don't reveal Assault from the Shadows card (#2312)
When a player had the ability cancellation option turned on and their
opponent used Assault from the Shadows to put a card into shadows, that
card was displayed on the prompt. This is because the prompt displays
any targets for the ability regardless of visibility. Most of the time
this is useful because the card must be revealed even while still in
hand to verify it meets the requirements for the ability (e.g. Core
Arianne Martell). In the case of Assault from the Shadows, there are no
characteristic requirements that need to be verified and thus the card
should not be revealed.

Thus, even though Assault from the Shadows uses the word "choose" and
thus the card should be considered a target, the implementation has been
changed to a non-targeting effect. This should prevent the card from
being revealed when the opponent is prompted to cancel and should have
minimal side effects regarding targeting.