Apply attachment lasting effect to specific parent
Previously, Heartbane and Crown of Golden Roses were generating a
lasting effect that applied to any card it was attached on, instead of
only the card it was attach on when the ability triggered. This led to a
bug where if the attachment was moved to another character, the new
character would improperly gain the effect applied to its original
parent.