Add support for blood and shadow tokens (#47)
* Add support for blood and shadow tokens

Adds the new token types required for Sea of Blood, Assault from the
Shadows, and Banished from the Light. Additionally, it allows tokens to
be displayed on facedown tokens, which is necessary for tokens on cards
in the shadows.

* Only fade tokens from attached cards

Previously, tokens were marked as faded if they were on an attachment
card. However, this caused shadow tokens on attachment cards in shadows
to be improperly faded. The logic has been updated to only apply the
`fade` property to cards that have been attached to the current card.

* Avoid overlapping token initials