Implement Reckless (#2339)
* Extract choosing defenders to separate prompt

* Add support for `mustSelect` to select card prompt

An array of cards may be sent using the `mustSelect` property to any
select card prompt. Cards in that array will be auto-selected if under
the limit for the prompt and then unselectable. If there are more
must-select cards than the prompt allows, the player must choose one of
the cards in the array.

* Use ChooseParticipantsPrompt for attackers

* Implement Reckless