Only reveal your own draw deck on searches
Previously, full deck searches would reveal all cards in all draw decks,
regardless of owner. This was hidden by the fact that the top of
opponent draw decks always displayed as a facedown card. However, the
implementation of Stolen Message changed this behavior to display the
card if it was visible, revealing the bug.

Now, card visibility for full deck searches is explicitly limited to
cards in your own draw deck.