Track applied targets separately for effects
Previously, a card would not be added as a target to an effect if it
could not be currently applied to the card. This led to difficulties
when a lasting effect was applied to a partially immune target. In these
cases, the effect should get applied after immunity is lost. However, in
cases where a card leaves play, the effect should be permanently
removed.

In order to satisfy both of these, the targets and applied targets of an
effect are now tracked separately. A card is a target of an effect if
matches the targeting characteristics of the effect. It only becomes an
applied target if the card is not immune to the effect. By tracking them
separately, targets can be permanently removed from lasting effects when
they leave play, and they can be applied to a non-removed target if
immunity is lost without improperly re-adding the target to the lasting
effect.