Allow Rickon Stark to cancel any "search" ability (#2399)
Previously, Rickon Stark was implemented to cancel searches directly.
However, per the card text, Rickon should cancel the effects of any
ability that has the word "search" in it. This includes cards where the
search is in a "Then" clause (Heir to the Iron Throne, Sea of Blood),
cards where the search is a choice (Dragonbinder), or even Rickon
himself.