Allow "pay X gold" prompts to have a 0 maximum (#2421)
Previously, if a card that had an X cost whose value could only be 0,
the prompt would be skipped without setting the `xValue` on the context
object. This lead to the player's gold becoming `NaN` for the rest of
the game with no way to recover. The most common scenario was playing
Gifts for the Widow while having no gold - you'd want to search for a 0
cost attachment but it would trigger the bug.

Now, the XValuePrompt will properly ask the player to choose the number,
even if it's only 0. As a safety measure, if the prompt is called with
an invalid range, it automatically resolve for the lower end of the
range to ensure an `xValue` is always set. While this would still be a
bug, it would be less catastrophic than rendering gold useless the rest
of the game.