Do not allow Duel to kneel knelt targets
Previously, Duel allowed you to choose an already knelt character to
kneel, or even a character that couldn't be knelt. Now you must
explicitly choose a standing, kneelable character.

Additionally, the prompt has been changed to a select prompt to be more
in line with the rest of the UI.