Change Risen / BMR text to lasting effect
Previously, the +1 STR effect on Risen from the Sea and the blanking
effect on Blood Magic Ritual were implemented as persistence effects.
This resulted in losing the effect when blanked by Brother's Robes,
which should only blank printed text and not gained text.

Now these effects are implemented as lasting effects that are active
while the cards remain in play.