Handle Tricksy Bird vs moving plots
Previously, if the plot selected by Tricksy Bird was moved out of the
plot deck by another ability, such as Ser Gerris Drinkwater or Change of
Plans, it would be selected in the following phase from the used pile.

Now, plot selection ensures that the "must reveal" plot is in the plot
deck. Additionally, if you have multiple copies of the same plot, one is
chosen for Tricksy, then moved out, you now properly have to reveal the
other copy instead.