Use `context.player` on Gate locations
Because the Gate cards from At the Gates sacrifice themselves as a cost,
they *must* refer to `context.player` instead of `this.controller`
because the controller does not reflect the player that controlled the
card before sacrifice.

Also pass `context` to `condition` methods to further reduce the usages
of `this.controller`.