Filter out null target selections
In order to support "if able" and "each player" target selections, the
value of the selection now has to allow null. The ability cancellation
UI tries to display a list of chosen cards, but was crashing due to null
values being in the array.

Now, null selections and unresolved selections are filtered out of that
array.

Fixes THRONETEKI-2EV
Fixes THRONETEKI-2EZ