Fix Gates of the Moon w/ Riddle + AtG
Because Varys's Riddle is implemented to temporarily "take control" of
the plot, when At the Gates is copied and brings out a card that applies
an effect to opponent plot card (such as Gates of the Moon), the current
controller is incorrect and the effect isn't applied. Since Riddle
wasn't raising the `onCardTakenControl` event, the effect engine never
knew to recalculate the effect.