Don't reveal hand early for Ghost of High Heart (#2548)
* Implement rudimentary 'Then' abilities

* Don't reveal hand early for Ghost of High Heart

Previously, the card selection and hand reveal for Ghost of High Heart
was being handled like a normal targeting prompt. However, this allows
the player to look at the opponent's hand before the opponent has a
chance to cancel the ability.

Now, the look-at-hand is implemented as the effect of the card, allowing
cancel prior to looking.