Fix "does not kneel" lasting effects (#2599)
* Fix 'does not kneel' lasting effects

Previously, "does not kneel as attacker / defender" lasting effects
would wear off after the first challenge they applied to. This was
because the challenge type was encoded as a condition, and when that
condition became false it cancelled the effect.

Now, where possible, the effect is no longer conditional. Instead, it
applies a separate effect type for each challenge.

* Properly check defense "does not kneel"