Do not execute cancelled simultaneous events (#2604)
* Do not execute cancelled simultaneous events

This should fix the issue with dupes not properly saving vs Poltical
Disaster, Varys (Core), and other simultaneous discards.

The SimultaneousEvents class groups together a series of events for
simultaneous execution. However, if one of those events is cancelled, it
should not be executed. The Event class itself achieves this by marking
a `parent` event object and notifying the parent to remove the event
upon cancellation. This mechanism is not directly possible with
simultaneous events because each event may have its own parent already.

Now, the SimultaneousEvents class will manually filter through cancelled
child events to ensure that they are properly excluded.

* Fix typo