Implement Maester Gormon + KL ruling (#2618)
* Reimplement King's Landing w/ a tracker

Uses a standard event tracker to implement King's Landing instead of
manually listening to events. The post-event calculation method needed
to be called in an additional spot in the event pipeline due to how the
events are currently organized. Otherwise, it was allowing two marshals
from King's Landing before de-activating the effect.

Additionally, updated the way marshalling into shadows works to match
the ruling that King's Landing does not allow a card to be marshalled
into shadows.

* Add original location for played card events

* Add explicit reveal top card flag

* Implement Maester Gormon