Allow players to look at the top of their deck (#70)
Changes how the draw deck behaves on the front end to allow players to
look at the top of their deck for Maester Gormon, without it looking
like the top card is revealed. Instead of a menu to open the draw deck,
it works like any other pile, but all cards are facedown by default.
Cards like Stolen Message now explicitly send data to reveal the top
card of the deck.