Make "considered attacking" state dependent
Keep track of declared attackers to ensure that Azor Ahai does not
unintentionally remove a declared Azor if the other R'hllor character is
removed.

By making `consideredToBeAttacking` a state dependent effect, it will
continually be re-evaluated. This will allow Azor Ahai to re-assert
itself in the event that a declared Azor Ahai is removed.