fix the azor ahai reborn bug (#2742)
* work in progress to fix the azor ahai reborn bug that stannis is still participating in a challenge even though the last rhllor character was removed

* added consideredToBeAttackingWasApplied flag to the character being added to a challenge via the "consideredToBeAttacking" effect to better handle the unapply method of the effect.

* Make "considered attacking" state dependent

Keep track of declared attackers to ensure that Azor Ahai does not
unintentionally remove a declared Azor if the other R'hllor character is
removed.

By making `consideredToBeAttacking` a state dependent effect, it will
continually be re-evaluated. This will allow Azor Ahai to re-assert
itself in the event that a declared Azor Ahai is removed.