Use replacement handler for Meryn Trant (#2851)
Because Ser Meryn Trant's ability is an interrupt and not a reaction,
and because it uses the word "instead", the ability needs to use a
replacement handler. Otherwise, the card will be moved out of the game
immediately, and then placed in the discard pile from the normal
onCardDiscarded event handler.