Add a prompt to handle naming cards (#2873)
Previously, card abilities that required you to name a card used basic
menu prompts and did checks on the cards manually. This introduces a
special prompt that makes it easier to limit the card named to given
`match` criteria, and better handles if the player does not name a card.