Remove attackers / defenders declared event
Previously, there were aggregate events for when attackers
and defenders were declared in a challenge. This was causing
cards that used the events like Motley and Breaker of Chains
to trigger for characters that were added to the challenge
via an effect (e.g. Azor Ahai).

These cards have been reimplemented to use the singular
events for declaring an attacker or defender (which will not
fire if added by an effect), and the aggregate event has
been removed.