Port end of game / rematch prompt from Keyteki (#2972)
* Copy over event + restricted list on rematch

* Add support for rematches (#53)

* Handle game end prompt when a draw occurs

* Disable the won prompt for testing

In our card integration tests, the players typically have very few
cards in their decks. This means they are immediately eliminated
from the game after setup, causing the game won prompt to pop up.

This adds a boolean flag for skipping that prompt during tests. We
can't directed set the winner or finishedAt to prevent it from
displaying because there are other tests that make assertions on
the winner and finishedAt times.