Fix Pulling the Strings + At the Gates interaction (#2999)
* Update usages of this.controller to context.player

This only updates the cards relevant to Pulling the Strings - any
card that is either a Edict, Kingdom, or Scheme. Additionally,
since City of Spiders is a Scheme and allows copying of City plots,
those have been updated as well.

* Allow Pulling the String to work w/ At the Gates

Instead of temporarily taking control of At the Gates (which
leads to it counting as a card in the player's used pile),
Pulling the Strings now initiates the ability directly with
the proper initiating player.