Fix interaction between Delena Florent and In Daznak's Pit (#3022)
* Remove legacy usages of Player.phase

* Fix trigger for Ranger's Cache

It should be an interrupt but was set as a reaction

* Ensure interrupts to end of phase occur in phase

Previously, the currentPhase was being reset before interrupts to
the onPhaseEnded event. Now the phase isn't reset until after any
interrupts. This should allow the effect for Delena Florent to be
active when In Daznak's Pit attempts to stand characters.