Fix the Knight gaining stealth
A previous fix aimed at Randyl Tarly and Lord of the Crossing
changed when a character was considered attacking. This broke
The Knight's ability to gain stealth before the challenge was
officially initiated.

The attacking / defending check has been reverted, and instead
Lord of the Crossing has been modified to only apply the
strength buff after the challenge has been initiated.