Automatically complete prompts for fake players
In single player games, we create a fake defending player during challenges
to avoid special casing every ability to handle a missing defending player.
This leads to a problem on abilities that prompt the defending player, such
as Late Summer Feast, since there's no player to respond to the prompt.

Now the menu and card select prompts will immediately complete themselves
if the player being prompted is the fake player. This may not cover every
case, but it should fix many of the existing hanging prompts.