Port private games from Keyteki (#3215)
* Update lobby code from keyteki

* Add back event-specific spectators for now

* Allow game to be marked as private

* Re-broadcast game list after user has authed

This fixes a bug where the game list isn't updated according to the
user's blocklist. The old flow is:
* First visit to the site, the initial list is sent as an anonymous
  user.
* User logs in / the existing login is validated
* List is not updated with any user-specific restrictions

* Add full flag to pending game

The existing client already obeys the full flag for certain things (e.g.
whether the Join button shows up), but checks player counts manually for
other things. Returning this data from the server instead of checking it
client side will be easier once there are variable player game such as
with Melee and Draft.

Co-authored-by: Stuart Walsh <stu@ipng.org.uk>