Fix Shagwell / Patchface loop (#3358)
* Fix feedback loop between Shagwell and Patchface

Previously, Shagwell and Patchface had independent effects that applied
dynamic keywords. However, if Patchface had a gold placed on it, it
became a valid source for Shagwell's effect. Because each effect was
independent, this led to Patchface keeping keywords even after the
original source left place or was no longer a source via Shagwell.

Example:
* Shagwell, Patchface, and Fickel Bannerman are in play, each with gold
  on them.
* Both Shagwell and Patchface receive Renown from Fickel Bannerman.
* Then Fickel Bannerman leaves play or has its gold removed.  Shagwell
  and Patchface inappropriately kept Renown - Shagwell was keeping it
  because Patchface still had it. Patchface was keeping it because
  Shagwell still had it.

Now, both these cards effect is to add a "keyword source" to the
receiving card. This keyword source is a function that returns true for
cards that match the effect criteria to provide the keywords (gold
tokens for Shagwell, Fool trait for Patchface).

These keyword source functions are then used in separate, global effect
that constructs a directed graph of cards that provide each other
keywords. The effect uses the graph to get the final list of connected
cards, get their keywords, and apply them to the card. This prevents
cycles between Shagwell and Patchface from inappropriately keeping
keywords, since the combined keywords for both of them is calculated and
added / removed at the same time.

* Updating Shagwell test

* Fix linting

---------

Co-authored-by: Stephen Patane <patane97@gmail.com>