Ensure game owner joins rematch games (#3362)
When we implemented game links last year, we added a "join" flag to the
newGame method. This wasn't being passed during rematch games, so the
owner of the game was no longer joining the rematch.

Also adds the game name to the rematch game, so we don't get games list
in the lobby as "undefined".