Fixing A Cask of Ale
Changed the event so you first have to select the two cards from the same controller and then which card you are moving the power from.
You are apparently supposed to be able to select two cards without power because it is a only referential target, but you can't play the event if there is no power in play.