Crone Forgotten Plans Interaction
I added the condition when selecting targets, this way you also can't play the event if there is no valid target. I also added this condition in Bloody Cup, because it now let's you select invalid targets on Crone turn. Not sure if this is okay solution for this, but if it is I can add it on some other events that I know also work like that.