Fix expiration condition for Last Greenseer
This makes a slight change to the custom duration effects to allow them
to listen on triggered versions of events (interrupts, reactions, etc)
instead of just the base event. This is definitely a hack, but is pretty
harmless and a non-invasive way to deal with the fact that base events
aren't currently being consistently raised if they occur as a result of
a thenAttachEvent.